keystate needs sending as long, not int.. fixes keystate corruption on net clients.
git-svn-id: https://g15daemon.svn.sourceforge.net/svnroot/g15daemon/trunk@322 ab83d0d9-5618-0410-a047-8efd564e0644

+++ g15daemon-wip/libg15daemon_client/g15daemon_client.h
@@ -67,7 +67,7 @@ int g15_send(int sock, char *buf, unsigned int len);
 int g15_recv(int sock, char *buf, unsigned int len);
 
 /* send a command (defined above) to the daemon.  any replies from the daemon are returned */
-int g15_send_cmd (int sock, unsigned char command, unsigned char value);
+unsigned long g15_send_cmd (int sock, unsigned char command, unsigned char value);
 /* receive an oob byte from the daemon, used internally by g15_send_cmd, but useful elsewhere */
 #define G15_FOREGROUND_SENT_OOB 1
 int g15_recv_oob_answer(int sock);
